Refactor point group sector organization and update imports
This commit introduces a new `sectors.py` module to encapsulate the definitions of various point group symmetry sectors, including `FiniteIrrepSector`, `SpinorIrrepSector`, `SpinfulPhaseSector`, `JointSpinfulPhaseSector`, and `SymmetryDegeneracy`. The existing imports in `__init__.py` and `ops.py` are updated to reflect this new structure, enhancing code organization and clarity. This change improves maintainability by separating sector definitions from operational logic.
